styling(icons): new icons for tabs and some component directory moving
