+++ src/common/components/Tabs/Tabs.styles.ts
@@ -1,9 +1,6 @@
-import * as React from 'react'
 import styled from 'styled-components'
-import { NavLink } from 'react-router-dom'
-import { MatchType } from '../../types/models'
 
-const TabsContainer = styled.div`
+export const TabsContainer = styled.div`
 	background:${/* eslint-disable-line */ props => props.theme.bg.gradientBlue};
 	border-radius:3px;
 	overflow: hidden;
@@ -59,37 +56,3 @@ const TabsContainer = styled.div`
 		padding:0 5px;
 	}
 `
-
-export interface Button {
-  linkClass?: string
-  iconClass: string
-  title?: string
-  path: string
-  state?: any
-}
-
-export interface Props {
-  buttons: Button[]
-  matchType: MatchType
-}
-
-export const Tabs: React.SFC<Props> = props => {
-  return (
-    <TabsContainer>
-      {props.buttons.map((button, index) => {
-        return (
-          <NavLink
-            className={button.linkClass ? button.linkClass : ''}
-            exact={props.matchType === MatchType.exact}
-            strict={props.matchType === MatchType.strict}
-            to={{ pathname: button.path }}
-            key={index}
-          >
-            {button.iconClass && <i className={button.iconClass} />}
-            {button.title && <p>{button.title}</p>}
-          </NavLink>
-        )
-      })}
-    </TabsContainer>
-  )
-}
