+++ src/components/project/ControlPanel/Connections/Connections.styles.ts
@@ -1,6 +1,6 @@
 import styled from 'styled-components'
 import { SquareButtonSection } from '../ControlPanel.styles'
-import { QRInner } from '../../../common/QRComponent'
+import { QRInner } from '../../../../common/components/QRCode/QRCode.styles'
 
 export const ConnectionButtonsWrapper = styled(SquareButtonSection)`
   button {
