+++ src/components/project/ProjectContainer.tsx
@@ -22,10 +22,10 @@ import { ProjectAgents } from './ProjectAgents'
 import { Spinner } from '../common/Spinner'
 import { UserInfo } from '../../modules/Account/types'
 import { ProjectSidebar } from './ProjectSidebar'
-import * as Toast from '../helpers/Toast'
+import * as Toast from '../../common/utils/Toast'
 import { deviceWidth } from '../../lib/commonData'
 import { ProjectClaimSubmitted } from './ProjectClaimSubmitted'
-import { explorerSocket } from '../helpers/explorerSocket'
+import { explorerSocket } from '../../common/utils/explorerSocket'
 import { FundingContainer } from './funding/FundingContainer'
 import { NotLedgered } from './overview/modalContent/NotLedgered'
