+++ src/components/project/curation/ProjectCreateUploadPublicDocs.tsx
@@ -7,7 +7,7 @@ import { decode as base64Decode, encode as base64Encode } from 'base-64'
 import { blankProjectData } from '../../../lib/commonData'
 import { Button, ButtonTypes } from '../../common/Buttons'
 import { FileLoader } from '../../common/FileLoader'
-import { successToast, errorToast } from '../../helpers/Toast'
+import { successToast, errorToast } from '../../../common/utils/Toast'
 import { CopyToClipboard } from 'react-copy-to-clipboard'
 
 const Text = styled.input`
