+++ src/components/project/curation/ProjectSignAndCreate.tsx
@@ -4,7 +4,7 @@ import { RootState } from '../../../common/redux/types'
 import { connect } from 'react-redux'
 import { decode as base64Decode } from 'base-64'
 import styled from 'styled-components'
-import { successToast, errorToast } from '../../helpers/Toast'
+import { successToast, errorToast } from '../../../common/utils/Toast'
 import { ErrorTypes } from '../../../types/models'
 import * as queryString from 'query-string'
 import { Banner } from './Banner'
