+++ src/components/project/funding/FundingContainer.tsx
@@ -3,13 +3,13 @@ import styled from 'styled-components'
 import { connect } from 'react-redux'
 import { RootState } from 'src/common/redux/types'
 import Web3Proxy from 'src/modules/web3/util/Web3Proxy'
-import * as Toast from '../../helpers/Toast'
+import * as Toast from '../../../common/utils/Toast'
 import { Web3Acc } from 'src/modules/web3/web3'
 import { FundingGauge } from './FundingGauge'
 import { FundingButton } from './FundingButton'
 import { Fragment } from 'react'
 import { ModalWrapper } from 'src/components/common/ModalWrapper'
-import { successToast, errorToast } from '../../helpers/Toast'
+import { successToast, errorToast } from '../../../common/utils/Toast'
 import { Button, ButtonTypes } from 'src/components/common/Buttons'
 import { BigNumber } from 'bignumber.js'
 import { Header } from '../../../types/models'
