+++ src/components/register/RegisterContainer.tsx
@@ -3,7 +3,7 @@ import styled from 'styled-components'
 import { UserInfo } from '../../modules/Account/types'
 import { connect } from 'react-redux'
 import { RootState } from '../../common/redux/types'
-import { errorToast, successToast } from '../helpers/Toast'
+import { errorToast, successToast } from '../../common/utils/Toast'
 import { ModalWrapper } from '../common/ModalWrapper'
 import { AgentRoles } from '../../types/models'
 import { Banner } from './Banner'
